# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Backend API for Work Hub — dashboard, DB manager, finances (PEN), and couple modules.",
        "title": "Work Hub API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api",
    "paths": {
...
        "/db/connect": {
            "post": {
                "security": [
                    {
                        "CookieAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "dbmanager"
                ],
                "summary": "Test a database connection",
                "parameters": [
                    {
                        "description": "Connection credentials",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dbmanager.connRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dbmanager.testConnectionResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httpx.APIError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/httpx.APIError"
                        }
                    }
                }
            }
        },
        "/db/query": {
            "post": {
                "security": [
                    {
                        "CookieAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "dbmanager"
                ],
                "summary": "Run a SQL query",
                "parameters": [
                    {
                        "description": "Connection credentials and SQL statement",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dbmanager.queryRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/db.QueryResult"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httpx.APIError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/httpx.APIError"
                        }
                    }
                }
            }
        },
        "/db/relationships": {
            "post": {
                "security": [
                    {
                        "CookieAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "dbmanager"
                ],
                "summary": "List foreign key relationships",
                "parameters": [
                    {
                        "description": "Connection credentials",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dbmanager.connRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dbmanager.listRelationshipsResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httpx.APIError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/httpx.APIError"
                        }
                    }
                }
            }
        },
        "/db/table/{name}/schema": {
            "post": {
                "security": [
                    {
                        "CookieAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "dbmanager"
                ],
                "summary": "Get table schema",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Table name",
                        "name": "name",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Connection credentials",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dbmanager.connRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dbmanager.tableSchemaResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httpx.APIError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/httpx.APIError"
                        }
                    }
                }
            }
        },
        "/db/tables": {
            "post": {
                "security": [
                    {
                        "CookieAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "dbmanager"
                ],
                "summary": "List tables",
                "parameters": [
                    {
                        "description": "Connection credentials",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dbmanager.connRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dbmanager.listTablesResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httpx.APIError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/httpx.APIError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "db.ColumnInfo": {
            "type": "object",
            "properties": {
                "default": {
                    "type": "string"
                },
                "isPrimaryKey": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "nullable": {
                    "type": "boolean"
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "db.ForeignKey": {
            "type": "object",
            "properties": {
                "fromColumn": {
                    "type": "string"
                },
                "fromTable": {
                    "type": "string"
                },
                "toColumn": {
                    "type": "string"
                },
                "toTable": {
                    "type": "string"
                }
            }
        },
        "db.QueryResult": {
            "type": "object",
            "properties": {
                "columns": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "executionTimeMs": {
                    "type": "integer"
                },
                "rowCount": {
                    "type": "integer"
                },
                "rows": {
                    "type": "array",
                    "items": {
                        "type": "object",
                        "additionalProperties": {}
                    }
                },
                "truncated": {
                    "type": "boolean"
                }
            }
        },
        "dbmanager.connRequest": {
            "type": "object",
            "properties": {
                "database": {
                    "type": "string"
                },
                "dialect": {
                    "type": "string"
                },
                "host": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "port": {
                    "type": "integer"
                },
                "user": {
                    "type": "string"
                }
            }
        },
        "dbmanager.listRelationshipsResponse": {
            "type": "object",
            "properties": {
                "relationships": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/db.ForeignKey"
                    }
                }
            }
        },
        "dbmanager.listTablesResponse": {
            "type": "object",
            "properties": {
                "tables": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "dbmanager.queryRequest": {
            "type": "object",
            "properties": {
                "database": {
                    "type": "string"
                },
                "dialect": {
                    "type": "string"
                },
                "host": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "port": {
                    "type": "integer"
                },
                "sql": {
                    "type": "string"
                },
                "user": {
                    "type": "string"
                }
            }
        },
        "dbmanager.tableSchemaResponse": {
            "type": "object",
            "properties": {
                "columns": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/db.ColumnInfo"
                    }
                },
                "table": {
                    "type": "string"
                }
            }
        },
        "dbmanager.testConnectionResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
        "httpx.APIError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Static token used by CLI and public share routes, sent as \"Bearer \u003ctoken\u003e\".",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "CookieAuth": {
            "description": "JWT access token issued by /api/auth/login, sent as the access_token cookie (not an Authorization header — see middleware.JWTAuth).",
            "type": "apiKey",
            "name": "access_token",
            "in": "cookie"
        }
    }
}